+++ lib/language_pack/helpers/nodebin.rb
@@ -1,17 +1,8 @@
 require 'json'
 
 class LanguagePack::Helpers::Nodebin
-  URL = "https://nodebin.herokai.com/v1/"
-
-  def self.query(q)
-    response = Net::HTTP.get_response(URI("#{URL}/#{q}"))
-    if response.code == '200'
-      JSON.parse(response.body)
-    end
-  end
-
   def self.hardcoded_node_lts
-    version = "10.14.1"
+    version = "10.15.3"
     {
       "number" => version,
       "url"    => "https://s3.amazonaws.com/heroku-nodebin/node/release/linux-x64/node-v#{version}-linux-x64.tar.gz"
@@ -19,22 +10,18 @@ class LanguagePack::Helpers::Nodebin
   end
 
   def self.hardcoded_yarn
-    version = "1.12.3"
+    version = "1.16.0"
     {
       "number" => version,
       "url"    => "https://s3.amazonaws.com/heroku-nodebin/yarn/release/yarn-v#{version}.tar.gz"
     }
   end
 
-  def self.node(q)
-    query("node/linux-x64/#{q}")
-  end
-
   def self.node_lts
-    hardcoded_node_lts # node("latest?range=6.x")
+    hardcoded_node_lts
   end
 
-  def self.yarn(q)
-    hardcoded_yarn # query("yarn/linux-x64/#{q}")
+  def self.yarn
+    hardcoded_yarn
   end
 end