+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Master
 
+* Set memory default for Node builds (https://github.com/heroku/heroku-buildpack-ruby/pull/861)
 * Default Ruby version is now 2.5.5, was previously 2.5.3 (https://github.com/heroku/heroku-buildpack-ruby/pull/863)
+* Default Node version is now 10.15.3 and default Yarn version is now 1.16.0 (https://github.com/heroku/heroku-buildpack-ruby/pull/884)
 
 ## v200 (3/7/2019)